refactor(backend): split telegram.py into chat_sync + chat_batch + bio_sync (1146 -> 299 lines)
